fixed lib and moves test to test file
