+++ lib/food_truck.rb
@@ -1,15 +1,1 @@
-require "minitest/autorun"
-require "minitest/pride"
-require "./lib/food_truck"
 
-class FoodTruckTest < Minitest::Test
-  def test_should_exist
-    food_truck = FoodTruck.new("Rocky Mountain Pies")
-    assert_instance_of FoodTruck, food_truck
-  end
-
-  def test_should_have_attributes
-    food_truck = FoodTruck.new("Rocky Mountain Pies")
-    assert_equal "Rocky Mountain Pies", item2.name
-  end
-end
