+++ test/food_truck_test.rb
@@ -1,0 +1,15 @@
+require "minitest/autorun"
+require "minitest/pride"
+require "./lib/food_truck"
+
+class FoodTruckTest < Minitest::Test
+  def test_should_exist
+    food_truck = FoodTruck.new("Rocky Mountain Pies")
+    assert_instance_of FoodTruck, food_truck
+  end
+
+  def test_should_have_attributes
+    food_truck = FoodTruck.new("Rocky Mountain Pies")
+    assert_equal "Rocky Mountain Pies", item2.name
+  end
+end
